refactor: moved types and parse function to separate packages
