+++ goman/internal/types/item.go
@@ -1,0 +1,9 @@
+package types
+
+type Item struct {
+	Type        string
+	Name        string
+	Signature   string
+	Description string
+	Example     string
+}
