+++ goman/internal/types/module.go
@@ -1,0 +1,7 @@
+package types
+
+type Module struct {
+	Name        string
+	Uri         string
+	Description string
+}
